refactor: useSyncExternalStore 제거
